rename multi-ota folder name, remove docs folder
