+++ src/platform/ESP32/BUILD.gn
@@ -98,18 +98,18 @@ static_library("ESP32") {
   if (chip_enable_ota_requestor) {
     if (chip_enable_multi_ota_requestor) {
       sources += [
-        "multi-ota/AppImageProcessor.cpp",
-        "multi-ota/AppImageProcessor.h",
-        "multi-ota/MultiImageOTAProcessorImpl.cpp",
-        "multi-ota/MultiImageOTAProcessorImpl.h",
-        "multi-ota/MultiOTAImageHeader.cpp",
-        "multi-ota/MultiOTAImageHeader.h",
-        "multi-ota/SubImageProcessor.h",
+        "multi_ota/AppImageProcessor.cpp",
+        "multi_ota/AppImageProcessor.h",
+        "multi_ota/MultiImageOTAProcessorImpl.cpp",
+        "multi_ota/MultiImageOTAProcessorImpl.h",
+        "multi_ota/MultiOTAImageHeader.cpp",
+        "multi_ota/MultiOTAImageHeader.h",
+        "multi_ota/SubImageProcessor.h",
       ]
       if (chip_enable_encrypted_ota) {
         sources += [
-          "multi-ota/EncryptedOTAHelper.cpp",
-          "multi-ota/EncryptedOTAHelper.h",
+          "multi_ota/EncryptedOTAHelper.cpp",
+          "multi_ota/EncryptedOTAHelper.h",
         ]
       }
     } else {
