+++ src/platform/ESP32/multi_ota/AppImageProcessor.cpp
