+++ src/platform/ESP32/multi_ota/MultiOTAImageHeader.cpp
